gui/backend/requirements.txt
# Palo Alto Upgrade GUI - Backend Requirements
# Install with: pip install -r backend/requirements.txt

# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# WebSocket
websockets==12.0
starlette==0.27.0

# Database
sqlalchemy==2.0.23
aiosqlite==0.19.0

# YAML parsing
pyyaml==6.0.1

# Process management
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Logging
loguru==0.7.2

# Serialization
python-json-logger==2.0.7
